requirements.txt
pyaudio==0.2.11
google-cloud-speech
pydub==0.23.1
protobuf
python3-pyaudio

